refactor(version): move to dunder file
