+++ mdbenchmark/__init__.py
@@ -18,6 +18,7 @@
 # You should have received a copy of the GNU General Public License
 # along with MDBenchmark.  If not, see <http://www.gnu.org/licenses/>.
 from mdbenchmark.cli import cli
+from mdbenchmark.__version__ import VERSION as __version__
 
 
 if __name__ == "__main__":
